refactor: phone-viewmodel api 핸들러 extension으로 관리
